Add Stamp component to ProtectedPage and enhance Stamps.tsx UI
Stamped component has been added to the ProtectedPage, improving the display of user stamps after login. The Stamps.tsx file has been upgraded, replacing a simple paragraph of activeStampCount with an improved UI. A dynamic list of SVG icons has been added, with active stamps and needed stamps visually differentiated. The changes also resulted in a modification of the package-lock.json and package.json files due to the addition of @heroicons/react dependency.