Refactor authentication process and improve user interface
This update restructures the Login and Register components, enhancing usability by replacing specific elements with reusable components such as Form, EmailInput, and FormSubmitButton. Additionally, it introduces an image rendering component for the website logo. Changes in global.css improve the visibility and the user interface. The authentication process has been also updated to pass message and isError parameters through a redirect enhancing error handling and feedback to the user.