Refactor 'Admin' function and add new features
In addition to refactoring the 'Admin' function, additional features have been added such as 'getStampTimestamps' and 'getLoginCounts' functions to fetch and display the data in line charts. Now, the 'Admin' function also includes a form which can be used to refresh and display the data, enhancing the admin page functionality.