Add leaflet map and FAQ page
This update includes the implementation of a new Frequently Asked Questions (FAQ) page which includes a dynamically imported leaflet map. Also added a heroicons dependency and created a new Map component. Updated the Nav component to include a link to the FAQ page and updated both react and react-dom versions. The Map component offers a Leaflet map which is compatible with Webpack's CSS import.